Refactor task management: update task status handling, add getTasks functionality, and improve error handling

+++ todo list/backend/controllers/task.controller.js
@@ -4,7 +4,7 @@ import { ApiErrors } from "../utils/ApiErrors.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/AsyncHandler.js";
 
-// add task 
+// add task
 const addTask = asyncHandler(async (req, res) => {
   const { taskName, taskDescription } = req?.body;
   if (!taskName) {
@@ -29,31 +29,45 @@ const addTask = asyncHandler(async (req, res) => {
 
 // Change the status of the task
 const taskStatus = asyncHandler(async (req, res) => {
- try {
-     const { taskId } = req?.params;
-     const { isCompleted } = req?.body; 
+  try {
+    const { taskId } = req?.query;
+    const { isCompleted } = req?.body;
+    console.log(taskId, isCompleted, "task id and is completed");
+    if (!taskId) {
+      throw new ApiErrors(400, "taskid or flag is missing");
+    }
+    const a = await Task.findById(taskId)
 
-     const task = await Task.findOneAndUpdate(taskId, {
-        isCompleted:isCompleted
-     },{
-        new:true
-     })
-     return res
-     .status(200)
-     .json(new ApiResponse(200, task))
- } catch (error) {
-    throw new ApiErrors(400, "error while changing the status of the Task")
- }
-
+    console.log(a);
+    // a.isCompleted = 
+    const task = await Task.findByIdAndUpdate(taskId,{
+      isCompleted: isCompleted
+    },{
+      new:true
+    })
+    return res.status(200).json(new ApiResponse(200, task));
+  } catch (error) {
+    throw new ApiErrors(400, "error while changing the status of the Task");
+  }
 });
 
 // delete task
-const deleteTask = asyncHandler(async (req,res) => {
-  
-    const {taskId} = req?.params;
-    await Task.deleteOne({_id : taskId})
-    return res
-    .status(200)
-    .json(new ApiResponse(200, {}))
-})
-export { addTask, taskStatus , deleteTask};
+const deleteTask = asyncHandler(async (req, res) => {
+  const { taskId } = req?.query;
+
+  if (!taskId) {
+    throw new ApiErrors(400, "no task id found");
+  }
+  await Task.deleteOne({ _id: taskId });
+  return res.status(200).json(new ApiResponse(200, {}));
+});
+
+// get tasks
+const getTasks = asyncHandler(async (req, res) => {
+  const tasks = await Task.find({ user: req?.user._id });
+  if (!tasks) {
+    throw new ApiErrors(400, "error While fetcing the tasks");
+  }
+  return res.status(200).json(new ApiResponse(200, tasks));
+});
+export { addTask, taskStatus, deleteTask, getTasks };
